requirements.txt
pyyaml
python-Levenshtein